requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
transformers==4.37.2
torch==2.0.0
# CRITICAL FIX: Numpy 2.0 broke older Torch versions. We pin it to <2.
numpy<2
requests
python-dotenv

sqlalchemy
psycopg2-binary